Updated sync logic to fire on node selection
